+++ src/lib.rs
@@ -32,6 +32,7 @@ pub mod scanner_code;
 #[cfg(feature = "documents")]
 pub mod scanner_docs;
 pub(crate) mod scanner_filter;
+pub mod search;
 #[cfg(all(feature = "shells", any(unix, windows)))]
 pub mod shells;
 pub mod store;